pyproject.toml
[project]
name = "striatica"
version = "0.3.0"
description = "A geometric atlas for machine intelligence"
authors = [
    {name = "spacecat",email = "25212708+spacecat@users.noreply.gitlab.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    # === UMAP REPRODUCIBILITY CHAIN — exact pins ===
    # These versions produced the Feb 27 2026 dataset positions.
    # Changing ANY of these will produce different 3D embeddings
    # even with the same random_state. Do not widen these ranges.
    # If you must update, regenerate ALL data and visually verify.
    "numpy==2.4.2",
    "scipy==1.17.1",
    "scikit-learn==1.8.0",
    "umap-learn==0.5.11",
    "hdbscan==0.8.41",
    # === Other deps (safe to update) ===
    "requests>=2.32.5,<3.0.0",
    "huggingface-hub>=0.28.0,<1.0.0",
    "safetensors>=0.5.0,<1.0.0",
    "python-dotenv>=1.1.0,<2.0.0"
]

[project.scripts]
striat = "pipeline.cli:main"

[project.optional-dependencies]
ml = [
    "torch>=2.6.0; python_version >= '3.12' and python_version < '3.15'",
    "sae-lens>=6.37.0; python_version >= '3.12' and python_version < '3.15'",
    "transformer-lens>=2.12.0; python_version >= '3.12' and python_version < '3.15'",
    "transformers>=4.40.0,<5.0.0; python_version >= '3.12' and python_version < '3.15'"
]


[[tool.poetry.packages]]
include = "pipeline"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "ruff>=0.15.4,<0.16.0"
]
